pyproject.toml
[tool.poetry]
name = "product_api"
version = "0.1.0"
description = ""
authors = ["Dario Thornhill <dariothornhill@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1.4"
django-allauth = "0.43.0"
django-crispy-forms = "1.10.0"
django-debug-toolbar = "3.2"
whitenoise = "5.2.0"
django-safedelete = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
